Create remote controllable garage door.ino
This project was designed using gsm shield and arduino uno. You have the right to open and close the garage door by sending an SMS or calling the designated GSM number from your phone. And if you are one of those people with authority, you can decide which numbers can open and close the garage door.